requirements.txt
Django
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
Markdown==2.6.10
PyJWT==1.5.3
pytz==2017.3
sqlparse==0.2.4
gunicorn
django-heroku
python-dotenv
django-crispy-forms
django-rest-swagger